refactor(settings): move interfaces/settings.ts → lib/settings/types.ts
Co-locate settings types with the rest of the settings module.
Update all 45 import sites.
