+++ web/src/app/admin/billing/BillingDetailsView.tsx
@@ -38,7 +38,7 @@ import { formatDateShort } from "@/lib/dateUtils";
 import { humanReadableFormatShort } from "@opal/time";
 import { NEXT_PUBLIC_CLOUD_ENABLED } from "@/lib/constants";
 import { useSettings } from "@/lib/settings/hooks";
-import { Tier } from "@/interfaces/settings";
+import { Tier } from "@/lib/settings/types";
 import useUsers from "@/hooks/useUsers";
 
 // ----------------------------------------------------------------------------
