+++ web/src/refresh-pages/admin/UsersPage/InviteOnlyCard.tsx
@@ -8,7 +8,7 @@ import { Switch } from "@opal/components";
 import { useSettings } from "@/lib/settings/hooks";
 import { SWR_KEYS } from "@/lib/swr-keys";
 import { toast } from "@/hooks/useToast";
-import { Settings } from "@/interfaces/settings";
+import { Settings } from "@/lib/settings/types";
 import { updateAdminSettings } from "@/lib/settings/svc";
 
 export default function InviteOnlyCard() {
